Address github threads
