+++ ic-auction/src/api.rs
@@ -4,12 +4,15 @@ use ic_canister::{generate_exports, update, Canister};
 use ic_cdk::export::candid::Principal;
 use ic_helpers::metrics::Interval;
 
-use crate::error::{AuctionError, Result};
-use crate::{AuctionInfo, AuctionState, BiddingInfo};
+use crate::{
+    error::{AuctionError, Result},
+    state::{AuctionInfo, AuctionState, BiddingInfo},
+};
 
 pub trait Auction: Canister + Sized {
     // TODO [CPROD-1056]: Remove default implementation as the
     // user may forget to overwrite it
+    #[deprecated = "Default implementation of `auction_state` is deprecated. Please overwrite it."]
     fn auction_state(&self) -> Rc<RefCell<AuctionState>> {
         panic!("auction_state is unimplemented")
     }
