+++ ic-auction/src/lib.rs
@@ -1,8 +1,3 @@
 pub mod api;
-mod state;
-mod types;
-
 pub mod error;
-
-pub use self::state::*;
-pub use self::types::*;
+pub mod state;
