+++ ic-auction/src/state.rs
@@ -3,16 +3,52 @@ use std::collections::HashMap;
 use ic_canister::ic_kit::ic;
 use ic_cdk::export::candid::{CandidType, Deserialize, Principal};
 use ic_helpers::metrics::Interval;
+use ic_helpers::tokens::Tokens128;
 use ic_storage::IcStorage;
 
-use crate::{
-    error::{AuctionError, Result},
-    AuctionInfo, BiddingInfo, Cycles, Timestamp,
-};
+use crate::error::{AuctionError, Result};
 
 // Minimum bidding amount is required, for every update call costs cycles, and we want bidding
 // to add cycles rather then to decrease them. 1M is chosen as one ingress call costs 590K cycles.
 pub const MIN_BIDDING_AMOUNT: Cycles = 1_000_000;
+
+pub type Cycles = u64;
+pub type TxId = u64;
+pub type Timestamp = u64;
+
+#[derive(CandidType, Debug, Clone, Deserialize, PartialEq)]
+pub struct AuctionInfo {
+    pub auction_id: usize,
+    pub auction_time: Timestamp,
+    pub tokens_distributed: Tokens128,
+    pub cycles_collected: Cycles,
+    pub fee_ratio: f64,
+    pub first_transaction_id: TxId,
+    pub last_transaction_id: TxId,
+}
+
+/// Current information about upcoming auction and current cycle bids.
+#[derive(CandidType, Debug, Clone, Deserialize)]
+pub struct BiddingInfo {
+    /// Proportion of the transaction fees that will be distributed to the auction participants.
+    ///
+    /// The value of 1.0 means that all fees go to the auction, 0.0 means that all the fees go to
+    /// the canister owner.
+    pub fee_ratio: f64,
+
+    /// Timestamp of the last auction.
+    pub last_auction: Timestamp,
+
+    /// Period of performing auctions. Auction cannot be started before `last_auction + auction_period`
+    /// IC time.
+    pub auction_period: Timestamp,
+
+    /// Total cycles accumulated since the last auction.
+    pub total_cycles: Cycles,
+
+    /// The amount of cycles the caller bid for the upcoming auction.
+    pub caller_cycles: Cycles,
+}
 
 //------------------------------------------------------------------------------
 // Bidding state
