+++ ic-factory/src/api.rs
@@ -19,6 +19,7 @@ pub trait FactoryCanister: Canister + Sized + PreUpdate {
     // implements this trait.
     // TODO [CPROD-1056]: remove default implementation since the user may forget
     // to overwrite this method
+    #[deprecated = "Default implementation of `factory_state` is deprecated. Please overwrite it."]
     fn factory_state(&self) -> Rc<RefCell<FactoryState>> {
         panic!("factory_state is not implemented")
     }
